moved RowType.java to top level package
