+++ app/src/androidTest/java/com/peternaggschga/gwent/data/RowDaoUnitTest.java
@@ -7,6 +7,8 @@ import androidx.room.rxjava3.EmptyResultSetException;
 import androidx.test.core.app.ApplicationProvider;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 
+import com.peternaggschga.gwent.RowType;
+
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
