+++ app/src/androidTest/java/com/peternaggschga/gwent/data/UnitDaoUnitTest.java
@@ -8,6 +8,7 @@ import androidx.test.core.app.ApplicationProvider;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 
 import com.peternaggschga.gwent.Ability;
+import com.peternaggschga.gwent.RowType;
 
 import org.junit.After;
 import org.junit.Before;
