+++ app/src/main/java/com/peternaggschga/gwent/RowType.java
@@ -1,4 +1,4 @@
-package com.peternaggschga.gwent.data;
+package com.peternaggschga.gwent;
 
 /**
  * An enum listing the different types of attack rows.
