+++ app/src/main/java/com/peternaggschga/gwent/data/UnitDao.java
@@ -10,6 +10,7 @@ import androidx.room.OnConflictStrategy;
 import androidx.room.Query;
 
 import com.peternaggschga.gwent.Ability;
+import com.peternaggschga.gwent.RowType;
 
 import java.util.Collection;
 import java.util.List;
