+++ app/src/main/java/com/peternaggschga/gwent/data/UnitEntity.java
@@ -11,6 +11,7 @@ import androidx.room.ForeignKey;
 import androidx.room.PrimaryKey;
 
 import com.peternaggschga.gwent.Ability;
+import com.peternaggschga.gwent.RowType;
 import com.peternaggschga.gwent.domain.damage.DamageCalculator;
 
 /**
