+++ app/src/main/java/com/peternaggschga/gwent/data/UnitRepository.java
@@ -7,6 +7,7 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
 import com.peternaggschga.gwent.Ability;
+import com.peternaggschga.gwent.RowType;
 
 import java.util.Collection;
 import java.util.List;
