+++ app/src/main/java/com/peternaggschga/gwent/domain/cases/RowStateUseCase.java
@@ -2,7 +2,7 @@ package com.peternaggschga.gwent.domain.cases;
 
 import androidx.annotation.NonNull;
 
-import com.peternaggschga.gwent.data.RowType;
+import com.peternaggschga.gwent.RowType;
 import com.peternaggschga.gwent.data.UnitEntity;
 import com.peternaggschga.gwent.data.UnitRepository;
 import com.peternaggschga.gwent.domain.damage.DamageCalculator;
