+++ app/src/test/java/com/peternaggschga/gwent/data/UnitEntityUnitTest.java
@@ -9,6 +9,7 @@ import static org.mockito.Mockito.only;
 import static org.mockito.Mockito.verify;
 
 import com.peternaggschga.gwent.Ability;
+import com.peternaggschga.gwent.RowType;
 import com.peternaggschga.gwent.domain.damage.DamageCalculator;
 
 import org.junit.Test;
