+++ app/src/test/java/com/peternaggschga/gwent/domain/cases/BurnUnitsUseCaseUnitTest.java
@@ -8,7 +8,7 @@ import static org.mockito.Mockito.when;
 
 import androidx.annotation.NonNull;
 
-import com.peternaggschga.gwent.data.RowType;
+import com.peternaggschga.gwent.RowType;
 import com.peternaggschga.gwent.data.UnitEntity;
 import com.peternaggschga.gwent.data.UnitRepository;
 import com.peternaggschga.gwent.domain.damage.DamageCalculator;
